refactor : FragmentDto 위치 변경 (dispatch -> dispatch/dto)
